refactor(quadlet): split logging tests into their own file
The fields.rs file was already at 462 code lines; adding the two
log-rotation tests for #1354 pushed it to 524, over the 500 hard limit
and tripping the line-limit gate. Move both logging tests to a new
fields_logging.rs file in the same directory so the split is
self-contained (no module reorganization, no imports change).

fields.rs: 535 -> 473 lines (62 fewer).
fields_logging.rs: 0 -> 72 lines (new file).

Pure mechanical move; no semantic change to either test. Build,
lib tests, and binary tests all stay green.
